+++ pywal/backends/fast_colorthief.py
@@ -1,6 +1,7 @@
 """
 Generate a colorscheme using fast_colorthief.
 """
+
 import logging
 import sys
 
@@ -17,7 +18,7 @@ from .. import colors
 
 
 def gen_colors(img):
-    """ Ask backend to generate 16 colors. """
+    """Ask backend to generate 16 colors."""
     raw_colors = fast_colorthief.get_palette(img, 16)
 
     return [util.rgb_to_hex(color) for color in raw_colors]
@@ -27,6 +28,7 @@ def adjust(cols, light, cols16):
     """Create palette."""
     cols.sort(key=util.rgb_to_yiq)
     raw_colors = [*cols, *cols]
+    raw_colors[0] = util.darken_color(cols[0], 0.80)
 
     return colors.generic_adjust(raw_colors, light, cols16)