+++ pywal/backends/schemer2.py
@@ -1,6 +1,7 @@
 """
 Generate a colorscheme using Schemer2.
 """
+
 import logging
 import shutil
 import subprocess
@@ -20,6 +21,7 @@ def adjust(cols, light, cols16):
     """Create palette."""
     cols.sort(key=util.rgb_to_yiq)
     raw_colors = [*cols[8:], *cols[8:]]
+    raw_colors[0] = util.darken_color(cols[0], 0.80)
 
     return colors.generic_adjust(raw_colors, light, cols16)
 
@@ -31,5 +33,5 @@ def get(img, light=False, cols16=False):
         logging.error("Try another backend. (wal --backend)")
         sys.exit(1)
 
-    cols = [col.decode('UTF-8') for col in gen_colors(img)]
+    cols = [col.decode("UTF-8") for col in gen_colors(img)]
     return adjust(cols, light, cols16)