+++ pywal/backends/wal.py
@@ -1,6 +1,7 @@
 """
 Generate a colorscheme using imagemagick.
 """
+
 import logging
 import re
 import shutil
@@ -13,11 +14,27 @@ from .. import util
 
 def imagemagick(color_count, img, magick_command):
     """Call Imagemagick to generate a scheme."""
-    flags = ["-resize", "25%", "-colors", str(color_count),
-             "-unique-colors", "txt:-"]
+    flags = [
+        "-resize",
+        "25%",
+        "-colors",
+        str(color_count),
+        "-unique-colors",
+        "txt:-",
+    ]
     img += "[0]"
 
-    return subprocess.check_output([*magick_command, img, *flags]).splitlines()
+    try:
+        output = subprocess.check_output(
+            [*magick_command, img, *flags], stderr=subprocess.STDOUT
+        ).splitlines()
+    except subprocess.CalledProcessError as Err:
+        logging.error("Imagemagick error: %s", Err)
+        logging.error(
+            "IM 7 disables stdout by default, check the wiki for the fix."
+        )
+        sys.exit(1)
+    return output
 
 
 def has_im():
@@ -33,11 +50,7 @@ def has_im():
     sys.exit(1)
 
 
-def gen_colors(img):
-    """Format the output from imagemagick into a list
-       of hex colors."""
-    magick_command = has_im()
-
+def try_gen_in_range(img, magick_command):
     for i in range(0, 20, 1):
         raw_colors = imagemagick(16 + i, img, magick_command)
 
@@ -51,13 +64,33 @@ def gen_colors(img):
         else:
             logging.warning("Imagemagick couldn't generate a palette.")
             logging.warning("Trying a larger palette size %s", 16 + i)
+    return raw_colors
 
-    return [re.search("#.{6}", str(col)).group(0) for col in raw_colors[1:]]
+
+def gen_colors(img):
+    """Format the output from imagemagick into a list
+    of hex colors."""
+    magick_command = has_im()
+
+    raw_colors = try_gen_in_range(img, magick_command)
+
+    try:
+        out = [re.search("#.{6}", str(col)).group(0) for col in raw_colors[1:]]
+    except AttributeError:
+        if magick_command == ["magick", "convert"]:
+            logging.warning("magick convert failed, using only magick")
+            magick_command = ["magick"]
+            raw_colors = try_gen_in_range(img, magick_command)
+            out = [
+                re.search("#.{6}", str(col)).group(0) for col in raw_colors[1:]
+            ]
+
+    return out
 
 
 def adjust(cols, light, cols16):
     """Adjust the generated colors and store them in a dict that
-       we will later save in json format."""
+    we will later save in json format."""
     raw_colors = cols[:1] + cols[8:16] + cols[8:-1]
 
     return colors.generic_adjust(raw_colors, light, cols16)
@@ -66,4 +99,8 @@ def adjust(cols, light, cols16):
 def get(img, light=False, cols16=False):
     """Get colorscheme."""
     colors = gen_colors(img)
+    # it is possible we could have picked garbage data
+    garbage = "# Image"
+    if garbage in colors:
+        colors.remove(garbage)
     return adjust(colors, light, cols16)